+++ build.js
@@ -1,7 +1,7 @@
 const { build } = require('esbuild')
 
 build({
-  entryPoints: ['./src/tombraidermenu.ts'],
+  entryPoints: ['./src/TombRaiderMenu.ts'],
   outfile: './dist/tombraider.js-v0.0.1.min.js',
   minify: true,
   bundle: true,
